test: migrate androidTest sources from com.vrpirates.rookieonquest to com.vrhub
