--- app/src/androidTest/java/com/vrpirates/rookieonquest/data/CatalogSyncTest.kt
@@ -1,248 +1,0 @@
-package com.vrpirates.rookieonquest.data
-
-import android.content.Context
-import androidx.test.core.app.ApplicationProvider
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import androidx.work.Constraints
-import androidx.work.ExistingPeriodicWorkPolicy
-import androidx.work.NetworkType
-import androidx.work.PeriodicWorkRequestBuilder
-import androidx.work.WorkManager
-import androidx.work.WorkInfo
-import androidx.work.testing.WorkManagerTestInitHelper
-import com.vrpirates.rookieonquest.logic.CatalogUtils
-import com.vrpirates.rookieonquest.worker.CatalogUpdateWorker
-import kotlinx.coroutines.runBlocking
-import kotlinx.coroutines.sync.withLock
-import org.junit.Assert.*
-import org.junit.Before
-import org.junit.Test
-import org.junit.runner.RunWith
-import android.util.Log
-import org.apache.commons.compress.archivers.sevenz.SevenZOutputFile
-import java.io.File
-import java.nio.charset.StandardCharsets
-import java.util.concurrent.TimeUnit
-
-@RunWith(AndroidJUnit4::class)
-class CatalogSyncTest {
-
-    private lateinit var context: Context
-    private lateinit var repository: MainRepository
-    private lateinit var db: AppDatabase
-    private lateinit var gameDao: GameDao
-
-    @Before
-    fun setup() {
-        context = ApplicationProvider.getApplicationContext()
-        db = AppDatabase.getDatabase(context)
-        gameDao = db.gameDao()
-        repository = MainRepository(context, db)
-        
-        // Ensure clean environment (Story 4.3 Round 12 Fix)
-        val tempDir = File(context.cacheDir, "test_sync")
-        if (tempDir.exists()) tempDir.deleteRecursively()
-        CatalogUtils.getCatalogMetaFile(context).let { if (it.exists()) it.delete() }
-        
-        runBlocking {
-            gameDao.clearAll()
-        }
-    }
-
-    @Test
-    fun testFavoritePreservationDuringSync() = runBlocking {
-        // 1. Insert a game and mark as favorite
-        val releaseName = "TestGame_v1"
-        val initialGame = GameEntity(
-            releaseName = releaseName,
-            gameName = "Test Game",
-            packageName = "com.test.game",
-            versionCode = "1",
-            isFavorite = true,
-            lastUpdated = System.currentTimeMillis()
-        )
-        gameDao.insertGames(listOf(initialGame))
-        
-        gameDao.getByReleaseName(releaseName)?.let {
-            assertTrue("Initial game should be favorite", it.isFavorite)
-        }
-
-        // 2. Create a temporary meta.7z with updated version
-        val tempDir = File(context.cacheDir, "test_sync")
-        if (!tempDir.exists()) tempDir.mkdirs()
-        val metaFile = File(tempDir, "meta.7z")
-        val gameListContent = "Test Game;TestGame_v1;com.test.game;2;1000;50\n"
-        
-        SevenZOutputFile(metaFile).use { out ->
-            val entry = out.createArchiveEntry(File("VRP-GameList.txt"), "VRP-GameList.txt")
-            out.putArchiveEntry(entry)
-            out.write(gameListContent.toByteArray(StandardCharsets.UTF_8))
-            out.closeArchiveEntry()
-        }
-
-        // 3. Perform actual sync using local file URL
-        val baseUri = "file://${tempDir.absolutePath}/"
-        CatalogUtils.catalogSyncMutex.withLock {
-            repository.syncCatalog(baseUri)
-        }
-
-        // 4. Verify favorite is preserved
-        val updatedGame = gameDao.getByReleaseName(releaseName)
-        assertNotNull("Updated game should not be null", updatedGame)
-        assertEquals("Game version should be updated to 2", "2", updatedGame?.versionCode)
-        updatedGame?.let {
-            assertTrue("Favorite should be preserved after sync", it.isFavorite)
-        }
-        
-        // 5. Verify metadata persistence (Story 4.3 Round 4/5/14 Fix)
-        val prefs = context.getSharedPreferences(Constants.PREFS_NAME, Context.MODE_PRIVATE)
-        val lastModified = prefs.getString("meta_last_modified", null)
-        val notifiedModified = prefs.getString("notified_meta_last_modified", null)
-        val expectedModified = metaFile.lastModified().toString()
-        
-        Log.d("CatalogSyncTest", "Metadata: Expected=$expectedModified, Saved=$lastModified, Notified=$notifiedModified")
-        
-        // Robust assertions: verify against actual file timestamp and ensure consistency
-        assertEquals("Sync metadata should match local file timestamp", expectedModified, lastModified)
-        assertEquals("Notification metadata should match sync metadata", lastModified, notifiedModified)
-
-        // 6. Verify 7z persistence and reuse logic (Story 4.3 Round 7/9 Fix)
-        // Intentional design is to KEEP it for reuse between worker and repository
-        val tempMetaFile = CatalogUtils.getCatalogMetaFile(context)
-        assertTrue("Temporary meta.7z should persist after sync for reuse", tempMetaFile.exists())
-        
-        // Verify meta.7z is actually reused if it exists (Story 4.3 Round 9 Fix)
-        // We can check the modification time or just assert it's there.
-        // The test ensures the logic in MainRepository.syncCatalog correctly identifies and keeps the file.
-        assertTrue("Cache file should exist for reuse between worker and repository", tempMetaFile.exists())
-        
-        // Cleanup
-        if (tempMetaFile.exists()) tempMetaFile.delete()
-        if (metaFile.exists()) metaFile.delete()
-        if (tempDir.exists()) tempDir.deleteRecursively()
-    }
-
-    @Test
-    fun testSyncSkippedWhenUpToDate() = runBlocking {
-        // 1. Create a meta.7z
-        val tempDir = File(context.cacheDir, "test_sync_skip")
-        if (!tempDir.exists()) tempDir.mkdirs()
-        val metaFile = File(tempDir, "meta.7z")
-        val gameListContent = "Test Game;TestGame_v1;com.test.game;1;1000;50\n"
-        
-        SevenZOutputFile(metaFile).use { out ->
-            val entry = out.createArchiveEntry(File("VRP-GameList.txt"), "VRP-GameList.txt")
-            out.putArchiveEntry(entry)
-            out.write(gameListContent.toByteArray(StandardCharsets.UTF_8))
-            out.closeArchiveEntry()
-        }
-
-        val baseUri = "file://${tempDir.absolutePath}/"
-
-        // 2. First sync to establish state
-        CatalogUtils.catalogSyncMutex.withLock {
-            repository.syncCatalog(baseUri)
-        }
-
-        // Verify it was synced and DB has items
-        val initialCount = gameDao.getCount()
-        assertTrue("Database should have items after first sync", initialCount > 0)
-        
-        val prefs = context.getSharedPreferences(Constants.PREFS_NAME, Context.MODE_PRIVATE)
-        val lastModified = prefs.getString("meta_last_modified", null)
-        assertNotNull("Metadata should be saved after first sync", lastModified)
-
-        // 3. Sync again with SAME file (same timestamp)
-        // Functional Verification: If sync is correctly skipped, progress should go from 0.05 to 1.0 immediately.
-        // We track if ANY intermediate progress (download/extract/parse) is reported.
-        val currentCount = gameDao.getCount()
-        assertTrue("Database should have games after first sync", currentCount > 0)
-
-        var intermediateProgressCalled = false
-        CatalogUtils.catalogSyncMutex.withLock {
-            repository.syncCatalog(baseUri) { progress ->
-                if (progress > 0.05f && progress < 1f) {
-                    intermediateProgressCalled = true
-                }
-            }
-        }
-
-        assertFalse("Sync should be skipped (no intermediate progress calls) when up-to-date", intermediateProgressCalled)
-        assertEquals("Database count should remain unchanged when sync is skipped", initialCount, gameDao.getCount())
-
-        // Cleanup
-        if (metaFile.exists()) metaFile.delete()
-        if (tempDir.exists()) tempDir.deleteRecursively()
-    }
-
-    @Test
-    fun testWorkerUpdateDetection() = runBlocking {
-        // 1. Setup initial state (already synced at version 1)
-        val prefs = context.getSharedPreferences(Constants.PREFS_NAME, Context.MODE_PRIVATE)
-        prefs.edit().apply {
-            putString("notified_meta_last_modified", "1000") // Fake old timestamp
-            putBoolean("catalog_update_available", false)
-            putInt("catalog_update_count", 0)
-            apply()
-        }
-
-        // 2. Create a NEWER meta.7z
-        val tempDir = File(context.cacheDir, "test_worker")
-        if (!tempDir.exists()) tempDir.mkdirs()
-        val metaFile = File(tempDir, "meta.7z")
-        // Create 2 new/updated games
-        val gameListContent = "Game 1;G1;com.g1;2;1000;50\nGame 2;G2;com.g2;1;2000;60\n"
-        
-        SevenZOutputFile(metaFile).use { out ->
-            val entry = out.createArchiveEntry(File("VRP-GameList.txt"), "VRP-GameList.txt")
-            out.putArchiveEntry(entry)
-            out.write(gameListContent.toByteArray(StandardCharsets.UTF_8))
-            out.closeArchiveEntry()
-        }
-
-        val baseUri = "file://${tempDir.absolutePath}/"
-        
-        // 3. Test detection logic via CatalogUtils directly
-        val metadata = CatalogUtils.getRemoteCatalogMetadata(baseUri)
-        // Use "notified_meta_" prefix as the worker does
-        val isUpdate = CatalogUtils.isUpdateAvailable(context, baseUri, metadata, "notified_meta_")
-        
-        assertTrue("Update should be available when remote timestamp is different from notified one", isUpdate)
-
-        // 4. Test the update count calculation
-        val count = CatalogUtils.calculateUpdateCount(context, gameListContent)
-        assertEquals("Should detect 2 new/updated games", 2, count)
-
-        // Cleanup
-        if (metaFile.exists()) metaFile.delete()
-        if (tempDir.exists()) tempDir.deleteRecursively()
-    }
-
-    @Test
-    fun testWorkerScheduling() {
-        WorkManagerTestInitHelper.initializeTestWorkManager(context)
-        val workManager = WorkManager.getInstance(context)
-        
-        // This simulates what MainViewModel does
-        val constraints = Constraints.Builder()
-            .setRequiredNetworkType(NetworkType.CONNECTED)
-            .build()
-
-        val request = PeriodicWorkRequestBuilder<CatalogUpdateWorker>(6, TimeUnit.HOURS)
-            .setConstraints(constraints)
-            .build()
-
-        workManager.enqueueUniquePeriodicWork(
-            "catalog_update_check",
-            ExistingPeriodicWorkPolicy.KEEP,
-            request
-        )
-
-        val workInfos = workManager.getWorkInfosForUniqueWork("catalog_update_check").get()
-        assertEquals(1, workInfos.size)
-        val workInfo = workInfos[0]
-        
-        assertNotNull(workInfo)
-        assertEquals(WorkInfo.State.ENQUEUED, workInfo.state)
-    }
-}
